Add files via upload
Charles Joseph Minard was a pioneer of data visualization and cartography, changing maps and numbers into modes of presenting compelling stories to the public. In this map, Minard's 1869 map of the disastrous Napoleonic invasion of Russia is reimagined using modern cartographic techniques and visual standards, while maintaining the poignant anti-war message of the original work. -Marcus Helm